% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{add_geom_vlines}
\alias{add_geom_vlines}
\title{Adds vertical lines to a ggplot object.}
\usage{
add_geom_vlines(plot, df, xintercepts, linetype, colour)
}
\arguments{
\item{plot}{Ggplot object}

\item{df}{A data frame with relevant plot data. (If the plot is faceted, then df must be the subset of data for the corresponding facet)}

\item{xintercepts}{Numerical x axis values that control the position of the line}

\item{linetype}{Aesthetic passed on to layer()}

\item{colour}{Aesthetic passed on to layer()}
}
\value{
Plot with vertical lines.
}
\description{
Adds vertical lines to a ggplot object.
}
