% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyser.R
\docType{class}
\name{Analyser-class}
\alias{Analyser-class}
\alias{Analyser}
\title{Class which generates output for every requested analysis.}
\description{
Reference class which is instanciated with an analysis description
as well as a plot generating function. With the command setData() the plots
are generated at runtime.
}
\section{Fields}{

\describe{
\item{\code{description}}{Description of the analysis}

\item{\code{plot_fnc}}{Function that generates plots.}

\item{\code{plots}}{List of generated plots}

\item{\code{plot_data}}{List of data frames for each generated plot.}

\item{\code{params}}{List of user defined parameters (from yaml)}

\item{\code{plot_settings}}{List of user defined plot settings (from yaml)}

\item{\code{ana_name}}{String naming the instanced analyser}
}}


