% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_directories}
\alias{check_directories}
\title{Tests whether user-defined input and output directories and their paths exist.
The output directory should not yet exist, but the input directory should.}
\usage{
check_directories(input_dir, output_dir)
}
\arguments{
\item{input_dir}{Path of an existing directory.}

\item{output_dir}{A path whose base name exists but the destination directory does not.}
}
\value{
True, if it doesn't throw an error.
}
\description{
Tests whether user-defined input and output directories and their paths exist.
The output directory should not yet exist, but the input directory should.
}
