% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_yaml_file}
\alias{check_yaml_file}
\title{Tests whether user-given yaml file exists.
If it does not exist, then it is pointed out how it can be generated
with default values.}
\usage{
check_yaml_file(yaml_file = NULL)
}
\arguments{
\item{yaml_file}{Path of yaml file.}
}
\value{
True if no error occurs
}
\description{
Tests whether user-given yaml file exists.
If it does not exist, then it is pointed out how it can be generated
with default values.
}
