% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{create_result_dirs}
\alias{create_result_dirs}
\title{Create directories to save results.}
\usage{
create_result_dirs(directory, yaml_outs = list())
}
\arguments{
\item{directory}{Path to the location where the results containing directories are to be created.}

\item{yaml_outs}{A list containing all analyses whose results the user wants to have output.}
}
\description{
Create directories to save results.
}
