% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createYaml.R
\name{createYaml}
\alias{createYaml}
\title{Creates a yaml file storing the parameters that are used for the analysis of the NOSA software results
and returns these parameters.}
\usage{
createYaml(
  yc,
  dirs = list(),
  manipulate = list(),
  plot_settings = list(),
  outputs = list()
)
}
\arguments{
\item{yc}{A yaml class object created by YAMLClass$new().}

\item{dirs}{List with the paths to the input and result directories and the sheet names to be read in.}

\item{manipulate}{List of parameters for editing input data; if empty, it is filled with defaults.}

\item{plot_settings}{List of parameters for designing plots.}

\item{outputs}{List of analyser names and their parameter; if empty, it is filled with defaults.}
}
\value{
List of all parameters used for the analysis#'
}
\description{
Creates a yaml file storing the parameters that are used for the analysis of the NOSA software results
and returns these parameters.
}
