% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extract_factor}
\alias{extract_factor}
\title{Extract factors from given name list.}
\usage{
extract_factor(names, factors)
}
\arguments{
\item{names}{List with names to search for factors.}

\item{factors}{List with factors.}
}
\value{
List of found factors. If not found
}
\description{
Extract factors from given name list.
}
