% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extract_values_between_two_given_times}
\alias{extract_values_between_two_given_times}
\title{Takes a data frame with two or more columns, one named as 'Time'.
Extracts the values for all columns, that are between
(and include) 'from' and 'to' of the 'Time' column.
Throws an error if 'from' or 'to' is not present.}
\usage{
extract_values_between_two_given_times(df, from, to, analyser = "")
}
\arguments{
\item{df}{Data frame of two columns. One named as 'Time'}

\item{from}{First timepoint of extracted values}

\item{to}{Last timepoint of extracted values}

\item{analyser}{String for error output. optional}
}
\value{
named columns with extracted values
}
\description{
Takes a data frame with two or more columns, one named as 'Time'.
Extracts the values for all columns, that are between
(and include) 'from' and 'to' of the 'Time' column.
Throws an error if 'from' or 'to' is not present.
}
