% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{filter_between_two_given_times}
\alias{filter_between_two_given_times}
\title{Takes a data frame with two or more columns.
Subset the data frame, retaining all rows where the values
are between (and include) 'from' and 'to' of the specified column.
Throws an error if 'from' or 'to' is not present.}
\usage{
filter_between_two_given_times(df, col_name, from, to, analyser = "")
}
\arguments{
\item{df}{Data frame of two or more columns, at least one of them is numeric.}

\item{col_name}{String representing the name of the (numeric) column to filter on.}

\item{from}{First point in time of resulting data frame}

\item{to}{Last point in time of resulting data frame}

\item{analyser}{String for error output. optional}
}
\value{
Subset of the input data frame
}
\description{
Takes a data frame with two or more columns.
Subset the data frame, retaining all rows where the values
are between (and include) 'from' and 'to' of the specified column.
Throws an error if 'from' or 'to' is not present.
}
