% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_analyser_objects}
\alias{get_analyser_objects}
\title{Get all requested analyser from outputs parameter list}
\usage{
get_analyser_objects(out_params)
}
\arguments{
\item{params}{List of analyser to be requested as well as their user defined parameters}
}
\value{
List of analyser objects
}
\description{
Get all requested analyser from outputs parameter list
}
