% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_bool_for_columns_by_factor}
\alias{get_bool_for_columns_by_factor}
\title{Takes a list of columnnames as well as a list of strings to search for in columnnames.
Returns ist list that holds for each search string a named list containg booleans for
the present of the string.
If no search strings are available the returning list contains of TRUEs except for
the excluded_column.}
\usage{
get_bool_for_columns_by_factor(
  columnnames,
  factors = NULL,
  excluded_column = FALSE
)
}
\arguments{
\item{columnnames}{List of strings.}

\item{factors}{List of search strings.}

\item{excluded_column}{String}
}
\value{
List that holds for each search string a named list containg booleans for
the present of the string.
}
\description{
Takes a list of columnnames as well as a list of strings to search for in columnnames.
Returns ist list that holds for each search string a named list containg booleans for
the present of the string.
If no search strings are available the returning list contains of TRUEs except for
the excluded_column.
}
