% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_boxplot}
\alias{get_boxplot}
\title{Function to create a boxplot}
\usage{
get_boxplot(df, factor, var, group = NULL)
}
\arguments{
\item{df}{Data frame.}

\item{factor}{Column name of x axis values.}

\item{var}{Column name of y axis values.}

\item{group}{Column name of factors that are connected with a line. Optional}
}
\value{
List of ggplot2 data
}
\description{
Function to create a boxplot
}
