% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_columns_by_factor}
\alias{get_columns_by_factor}
\title{Extract all columns that contain a given string (a factor) and return new data.frame.
If no factor is provided, all data are returned.}
\usage{
get_columns_by_factor(data, factors, needs_time = FALSE)
}
\arguments{
\item{data}{Data frame.}

\item{factors}{List with strings that are searched for in the column names of a given data.frame.}
}
\value{
Data.frame that consists only of the columns that contain the factors you are looking for in their names.
}
\description{
Extract all columns that contain a given string (a factor) and return new data.frame.
If no factor is provided, all data are returned.
}
