% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_columns_by_key}
\alias{get_columns_by_key}
\title{Extract all columns that contain a given string and return new data.frame.
If no list of strings is provided, all data are returned.}
\usage{
get_columns_by_key(data, substr_colnames, include_col = NULL)
}
\arguments{
\item{data}{Data frame.}

\item{substr_colnames}{List with strings that are searched for in the column names of a given data.frame.}

\item{include_col}{String for an additional colum. Optional.}
}
\value{
Data.frame that consists only of the columns that contain the substr_colnames you are looking for in their names.
}
\description{
Extract all columns that contain a given string and return new data.frame.
If no list of strings is provided, all data are returned.
}
