% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_factor_df}
\alias{get_factor_df}
\title{From a given list of column names extract substrings and return a new data.frame containing a column with substring reduced names
and a second column containing the extracted substrings}
\usage{
get_factor_df(names, factors)
}
\arguments{
\item{names}{List of Strings}

\item{factors}{List of (sub-)Strings}
}
\value{
Data.frame: first column holds substring reduced names, second column holds extracted substrings
}
\description{
From a given list of column names extract substrings and return a new data.frame containing a column with substring reduced names
and a second column containing the extracted substrings
}
