% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_plot_data}
\alias{get_plot_data}
\title{Extract data computed by ggpubr functions (e.g. mean_se)}
\usage{
get_plot_data(plot)
}
\arguments{
\item{plot}{Plot generated with ggpubr and additional function as mean_se}
}
\value{
Data frame with columns Time(x-values), Mean(y-values), SEM
}
\description{
Extract data computed by ggpubr functions (e.g. mean_se)
}
