% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_times_of_max_in_window}
\alias{get_times_of_max_in_window}
\title{Takes a data frame with two or more columns.
Subsets the data frame, retaining all rows where the values
are between (and include) 'from' and 'to' of the specified column.
For each column finds the maximum value and returns its point of time.}
\usage{
get_times_of_max_in_window(df, start, end, time_col_name)
}
\arguments{
\item{df}{Data frame containing the column "Time" as well as other columns in which the maximum value is to be found.}

\item{start}{Double that defines the start time}

\item{end}{Double that defines the end time}

\item{time_col_name}{String containing the column name of time.}
}
\value{
Named list: Names are column names from input data frame; values is the point of time, where the max value was found
}
\description{
Takes a data frame with two or more columns.
Subsets the data frame, retaining all rows where the values
are between (and include) 'from' and 'to' of the specified column.
For each column finds the maximum value and returns its point of time.
}
