% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{get_traceplot}
\alias{get_traceplot}
\title{Function to create a trace plot. Plotting the standard error of mean is optional.}
\usage{
get_traceplot(df, x_values, y_values, xlab, ylab)
}
\arguments{
\item{df}{Data frame.}

\item{x_values}{Column name of x axis values.}

\item{y_values}{Column name of y axis values.}

\item{xlab}{String title of x axis.}

\item{ylab}{String title of y axis.}
}
\value{
List of ggplot2 data
}
\description{
Function to create a trace plot. Plotting the standard error of mean is optional.
}
