% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getOutputFilenames}
\alias{getOutputFilenames}
\title{A list of file names for every output that is created during the analysis.}
\usage{
getOutputFilenames(output_dir, r_data_out = FALSE)
}
\arguments{
\item{output_dir}{Path to the location where result files are to be stored.}

\item{r_data_out}{Boolean; Should the read-in Nosa result data, as well as all executed metrics, be saved as an R object?}
}
\description{
A list of file names for every output that is created during the analysis.
}
