% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{manipulate_data}
\alias{manipulate_data}
\title{Checks whether given parameters can be applied and edits the data accordingly if necessary:
* subset by GroupingKeyWords
* checks whether the data can be paired by name
* crops data frames containing a time column
* normalizes data frames containing a time column}
\usage{
manipulate_data(data, params, paired, ana_name)
}
\arguments{
\item{data}{A data frame.}

\item{params}{Named list of parameters to be tested.}

\item{paired}{Boolean indicating whether the data is paired data.}

\item{ana_name}{String indicating the name of the analyser for which the data is to be processed.}
}
\value{
List with data = manipulated data, skipping = boolean for success, paired = boolean for pairable data
}
\description{
Checks whether given parameters can be applied and edits the data accordingly if necessary:
* subset by GroupingKeyWords
* checks whether the data can be paired by name
* crops data frames containing a time column
* normalizes data frames containing a time column
}
