% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{normalize}
\alias{normalize}
\title{Subset a data frame by (a) matching a pattern in names and (b) keeping
all rows where the numeric 'Time' values are between (and include) 'From' and 'To'.
Form row means for each sample and subtract (absolute normalization) or
divide (relative normalization) this mean from all paired samples.}
\usage{
normalize(data, params, grouping_keys)
}
\arguments{
\item{data}{A data frame. One column must contain the keyword 'Time'.}

\item{params}{Named list of normalization parameter.}

\item{grouping_keys}{Vector or array of strings for paired sample detection.}
}
\value{
List with success=boolean, data=data frame with normalized data
}
\description{
Subset a data frame by (a) matching a pattern in names and (b) keeping
all rows where the numeric 'Time' values are between (and include) 'From' and 'To'.
Form row means for each sample and subtract (absolute normalization) or
divide (relative normalization) this mean from all paired samples.
}
