% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NosaResultLoader.R
\docType{class}
\name{NosaResultLoader-class}
\alias{NosaResultLoader-class}
\alias{NosaResultLoader}
\title{Class to read Nosa software results stored as *.xlsx file and store the tables internally.}
\description{
The 'data' field is initialized after $loadNosaResults was called.
}
\section{Fields}{

\describe{
\item{\code{data}}{NosaResults data as list. Valid list entries are: "metadata", "Raw", "Processed", "Baseline", "Spike Detection" and "Smoothing".}
}}


