% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_EachSample.R
\name{output_EachSample}
\alias{output_EachSample}
\title{Function to create the standard error of mean of a given dataset.
If stati is empty, all existing data is used for the analysis.}
\usage{
output_EachSample(data, output_dir)
}
\arguments{
\item{data}{Data frame containing all processed sample data.}

\item{output_dir}{Path to the location where the resulting plots are to be stored.}
}
\value{
List with sample plots
}
\description{
Function to create the standard error of mean of a given dataset.
If stati is empty, all existing data is used for the analysis.
}
