% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_PeakCount.R
\name{output_PeakCount}
\alias{output_PeakCount}
\title{Function to count Peaks within user defined time window and return resulting boxplot.}
\usage{
output_PeakCount(data, params, dir)
}
\arguments{
\item{data}{Data frame.}

\item{params}{List that holds yaml defined parameters}

\item{dir}{Path to the location where the resulting plots are to be stored.}
}
\value{
List with ggplot2 peak count boxplot
}
\description{
Function to count Peaks within user defined time window and return resulting boxplot.
}
