% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_Responses.R
\name{output_Responses}
\alias{output_Responses}
\title{Function to extract responses. The time of the stimulus and a time window are
defined by the user. The x-axis is grouped by factor. Additional grouping of
different stimuli is possible.}
\usage{
output_Responses(data, params, dir)
}
\arguments{
\item{data}{Data frame.}

\item{params}{List that holds yaml defined parameters}

\item{dir}{Path to the location where the resulting plots are to be stored.}
}
\value{
List with ggplot2 reesponses boxplot
}
\description{
Function to extract responses. The time of the stimulus and a time window are
defined by the user. The x-axis is grouped by factor. Additional grouping of
different stimuli is possible.
}
