% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_SEM.R
\name{output_SEM}
\alias{output_SEM}
\title{Function to create the standard error of mean of a given dataset.
If stati is empty, all existing data is used for the analysis.}
\usage{
output_SEM(data, status_col, stati, output_dir)
}
\arguments{
\item{data}{Data frame on which the SEM is to be determined.}

\item{status_col}{Contains the status for each sample. If a value is NA, then it is not taken into account.}

\item{stati}{List of stati that should be seperated from each other.}

\item{output_dir}{Path to the location where the resulting plots are to be stored.}
}
\value{
List with SEM plots
}
\description{
Function to create the standard error of mean of a given dataset.
If stati is empty, all existing data is used for the analysis.
}
