% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_Trace.R
\name{output_Trace}
\alias{output_Trace}
\title{Function that converts the time sequence for each entry per specified sheet into a plot.}
\usage{
output_Trace(data, sheetname, dir)
}
\arguments{
\item{data}{List containing all read data from nosa results.}

\item{dir}{Path to the location where the resulting plots are to be stored.}

\item{sheetnames}{List containing all sheetnames for this analysis.}
}
\value{
List with ggplot2 sample plots
}
\description{
Function that converts the time sequence for each entry per specified sheet into a plot.
}
