% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performAnalysis.R
\name{performAnalysis}
\alias{performAnalysis}
\title{Perform the analysis of the NOSA software results.}
\usage{
performAnalysis(directory = NULL, yaml_obj = list(), yaml_file = NULL)
}
\arguments{
\item{directory}{Path to nosa result files to bee analysed.}

\item{yaml_obj}{List of parameters; if it is empty, it is filled with defaults}
}
\value{
List with filenames
}
\description{
A yaml object can also be provided to define which analyses are to be performed.
}
