% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_line}
\alias{plot_line}
\title{Create a line plot from a data frame, where the column names are at least 'x', 'y'.
Additional column names must correspond to grouping variables specified in
'facet_by' or 'column_name'.}
\usage{
plot_line(
  df,
  add = "none",
  display = NULL,
  facet_by = NULL,
  color_column = "white",
  area_from = 0,
  area_to = 0
)
}
\arguments{
\item{df}{A data frame.}

\item{add}{Character vector for adding plot elements that will be calculated (e.g. 'mean_se' or 'mean_sd')}

\item{display}{Character vector for additional geom elements: 'area', 'ribbon', 'linerange'}

\item{facet_by}{Character vector naming the columns the plot should be faceted by.}

\item{color_column}{String matching a column name for the group by which the plot lines are to be coloured.}

\item{area_from}{Starting value of geom_area.}

\item{area_to}{End value of geom_area.}
}
\value{
List with 'plot' = lineplot and 'data' = plot data (including 'mean_se' if calculated)
}
\description{
Create a line plot from a data frame, where the column names are at least 'x', 'y'.
Additional column names must correspond to grouping variables specified in
'facet_by' or 'column_name'.
}
