% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{reduce_data_by_window}
\alias{reduce_data_by_window}
\title{At any given time, with a defined time period, the data is extracted and written separately into individual tables.}
\usage{
reduce_data_by_window(df, window, timepoints, larger_window = FALSE)
}
\arguments{
\item{df}{Data.frame}

\item{window}{List with two doubles. The first determines the range before, and the second the range after a specific point in time.}

\item{timepoints}{List with one ore more timepoints.}

\item{larger_window}{Optional boolean. If TRUE, the amount of data is extended by two seconds in each direction.#'}
}
\value{
List of data.frames.
}
\description{
At any given time, with a defined time period, the data is extracted and written separately into individual tables.
}
