% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rowSem}
\alias{rowSem}
\title{From row return standard error of mean for arrays.}
\usage{
rowSem(x)
}
\arguments{
\item{x}{An array containing numeric values.}
}
\value{
SEM of row.
}
\description{
From row return standard error of mean for arrays.
}
