% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{separate_data_by_key}
\alias{separate_data_by_key}
\title{A data frame is separated into named lists (keywords).
The data is extracted according to the key in the column name
and stored into individual data frames. With additional 'global_cols'
the result consists of the global columns and the key seperated
data frames.}
\usage{
separate_data_by_key(df, keys, global_cols = NULL)
}
\arguments{
\item{df}{Data.frame}

\item{keys}{List of Strings}

\item{global_cols}{List of column names to be part of the result}
}
\value{
List of data.frames.
}
\description{
A data frame is separated into named lists (keywords).
The data is extracted according to the key in the column name
and stored into individual data frames. With additional 'global_cols'
the result consists of the global columns and the key seperated
data frames.
}
