% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{subset_by_key}
\alias{subset_by_key}
\title{Subset a data frame by matching patterns in their names.
If no list of patterns is provided, all data are returned.}
\usage{
subset_by_key(data, substr_colnames, include_col = NA_character_)
}
\arguments{
\item{data}{Data frame.}

\item{substr_colnames}{List with strings that are searched for in the column names.}

\item{include_col}{String for an additional column. Optional.}
}
\value{
Data.frame, which consists only of the columns containing the substr_colnames and the include_col in their names.
}
\description{
Subset a data frame by matching patterns in their names.
If no list of patterns is provided, all data are returned.
}
