% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YamlClass.R
\docType{class}
\name{YamlClass-class}
\alias{YamlClass-class}
\alias{YamlClass}
\title{Yaml object for specific parameters.}
\description{
If the object has the param, then return it.
If the param is unknown, create it with the given default value and return the default.
}
\section{Fields}{

\describe{
\item{\code{yamlObj}}{A Yaml object}
}}

\section{Methods}{

\describe{
\item{\code{getYaml(param_name, default)}}{Request a specific parameter and return its value if it exists. If it does not exist it is created with a default value.}

\item{\code{writeYaml(filename)}}{Write the YAML config to a YAML file. Returns TRUE on success (always), unless writing the file generates an error.}
}}

