% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{adjust_facet_width_of_plot}
\alias{adjust_facet_width_of_plot}
\title{Converts a ggplot object to grob object, extracts the amount of unique x-axis values per facet,
and change the relative width of the facet columns according to the amount.}
\usage{
adjust_facet_width_of_plot(ggplot, data)
}
\arguments{
\item{ggplot}{Facetted ggplot object}

\item{data}{sldlds}
}
\value{
ggplot object with (according to amount of x-axis values per facet) adjusted widths
}
\description{
Converts a ggplot object to grob object, extracts the amount of unique x-axis values per facet,
and change the relative width of the facet columns according to the amount.
}
