% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyser.R
\docType{class}
\name{Analyser-class}
\alias{Analyser-class}
\alias{Analyser}
\title{Class which generates output for every requested analysis.}
\description{
Reference class which is instanciated with an analysis description
as well as a plot generating function. With the command setData() the plots
are generated at runtime.
}
\section{Fields}{

\describe{
\item{\code{description}}{Description of the analysis}

\item{\code{plot_fnc}}{Function that generates plots.}

\item{\code{plots}}{List of plots}

\item{\code{dir_name}}{Name of the dir in which the plots are stored.}
}}


