% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extract_key}
\alias{extract_key}
\title{Extract key words from given name list.}
\usage{
extract_key(names, keys)
}
\arguments{
\item{names}{List with names to search for keys.}

\item{keys}{List with keys.}
}
\value{
Vector of found keys. If not found, entry is 'NA'.
}
\description{
Extract key words from given name list.
}
