% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_analyser_objects}
\alias{get_analyser_objects}
\title{Get all requested analyser from outputs parameter list}
\usage{
get_analyser_objects(params, statistics)
}
\arguments{
\item{params}{List of analyser to be requested as well as their user defined parameters}

\item{statistics}{Named list: paired is of type boolean (TRUE samples are paired samples), method is String that names the method to use (t.test, wilcox.test).}
}
\value{
List of analyser objects
}
\description{
Get all requested analyser from outputs parameter list
}
