% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_SEM.R
\name{get_SEM_plot}
\alias{get_SEM_plot}
\title{Function to create the plot of standard error of mean of a given dataset.}
\usage{
get_SEM_plot(df, columns)
}
\arguments{
\item{df}{Data frame on which the SEM is to be determined.}

\item{columns}{Vector of booleans to select columns for analysis.}
}
\value{
List of ggplot2 data
}
\description{
Function to create the plot of standard error of mean of a given dataset.
}
