% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_times_of_max_in_window}
\alias{get_times_of_max_in_window}
\title{For each column find the max value and return the timepoint.}
\usage{
get_times_of_max_in_window(df, start, end, time_col_name)
}
\arguments{
\item{df}{Data frame containing the column "Time" as well as other columns in which the maximum value is to be found.}

\item{start}{Double that defines the start time}

\item{end}{Double that defines the end time}

\item{time_col_name}{String containing the column name of time.}
}
\value{
Named list. Names were column names from input data frame; values is the timepoint, where the max value was found
}
\description{
For each column find the max value and return the timepoint.
}
