% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_times_of_max_in_window}
\alias{get_times_of_max_in_window}
\title{For each column find the max value and return the timepoint.}
\usage{
get_times_of_max_in_window(df, stim, window)
}
\arguments{
\item{df}{Data frame containing the column "Time" as well as other columns in which the maximum value is to be found.}

\item{stim}{Double representing the start time.}

\item{window}{Double that defines the end time (stim+window)}
}
\value{
Named list. Names were column names from input data frame; values is the timepoint, where the max value was found
}
\description{
For each column find the max value and return the timepoint.
}
