% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_traceplot}
\alias{get_traceplot}
\title{Function to create a trace plot. Plotting the standard error of mean is optional.}
\usage{
get_traceplot(df, var1, var2, sem = NULL, auc = NULL)
}
\arguments{
\item{df}{Data frame.}

\item{var1}{Column name of x axis values.}

\item{var2}{Column name of y axis values.}

\item{sem}{Column name of standard error of mean values. Optional.}

\item{auc}{Column name of auc booleans. Optional.}
}
\value{
List of ggplot2 data
}
\description{
Function to create a trace plot. Plotting the standard error of mean is optional.
}
