% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performAnalysis.R
\name{performAnalysis}
\alias{performAnalysis}
\title{Perform the analysis of NOSA software results.}
\usage{
performAnalysis(yaml_file = character())
}
\arguments{
\item{yaml_file}{A character string that provides the path and name of a yaml_file containing
the configuration parameters.
The function 'writeDefaultYaml(filename)' can be used to create
this file filled with all default parameters and values.}
}
\value{
List with sections data and plots.
}
\description{
By calling this function, all data specified under "Sheets" in the loaded Yaml
file are read into data.frames. Additional parameters and plots can be defined
with the Yaml configuration. It is advisable to create a file with the function
"writeDefaultYaml(filename)" and to adjust or deactivate the desired metrics.
}
