% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{reduce_data_by_factor}
\alias{reduce_data_by_factor}
\title{At any given time, with a defined time period, the data is extracted and written separately into individual tables.}
\usage{
reduce_data_by_factor(df, factors)
}
\arguments{
\item{df}{Data.frame}

\item{factors}{List of Strings.}
}
\value{
List of data.frames.
}
\description{
At any given time, with a defined time period, the data is extracted and written separately into individual tables.
}
