% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDefaultYaml.R
\name{writeDefaultYaml}
\alias{writeDefaultYaml}
\title{Write a yaml file with all possible default values for the nosa software analysis.}
\usage{
writeDefaultYaml(filename)
}
\arguments{
\item{filename}{a character string naming a file for writing}
}
\description{
It is recommended to make a copy and then modify it according to the
requirements of the analysis.
}
